feat(contacts): unify desktop and mobile contacts into one responsive page
One standalone responsive contacts component now serves both the desktop
(/contacts) and mobile-wallet (/wallet/contacts) routes — a flex table at
wide widths that reflows to stacked cards below 600px (pure CSS @media, no
BreakpointObserver, no desktop/mobile fork), mirroring the mining dashboard.

- Data source is ContactsStoreService only; the old inline localStorage
  Contact interface, load/save and legacy migration are removed.
- Network resolves per mode: NodeService.isRemote() (desktop remote + all
  nodeless) reads BtcxWalletService.network(), Core reads NgRx selectNetwork.
- Search (name/address/notes) feeds the FitRowsDirective + mat-paginator
  fit-to-viewport pagination, now applied on desktop too (all widths).
- Full wrapping address (word-break) in the card layout; finer address
  validation keys (invalid_format/invalid_checksum).
- Mobile /wallet/contacts route points at the unified component; the mobile
  wallet-contacts.component.ts duplicate is deleted.
